Make leader_schedule a utitlity module named leader_schedule_utils
