+++ src/lib.rs
@@ -41,7 +41,7 @@ pub mod gen_keys;
 pub mod gossip_service;
 pub mod leader_confirmation_service;
 pub mod leader_schedule;
-pub mod leader_scheduler;
+pub mod leader_schedule_utils;
 pub mod local_vote_signer_service;
 pub mod packet;
 pub mod poh;
